Update tests to use renamed Unfurl core code.
